Move CSS from inline app.R to main.css.
Switched font sizes to use relative values instead of pixels. Made
sure that there is only one scrollbar visible in the dashboard.
Lightened the tick marks for the "number of words" slider, since
newer versions of the blue dashboard theme don't have the CSS rule
built in to make them stand out.

Also, added a title to the webpage, made sure that the custom CSS
can override Bootstrap's, and fixed a bug with the implementation
of the aforementioned tick marks.