Move `generateClusters()` to UI section.
Preparing an attempt at fixing the Clusters tab so that clusters
are not reset when the user changes the number of shown words.

Fix warning message "Input to `asJSON(keep_vec_names=TRUE)` is a
named vector. In a future version of `jsonlite`, this option will
not be supported". It appears to have occurred due a default value
for "options" in `DT::datatable()`.