Don't use Datatables paging for Operations tables.
Changing the page length didn't do anything before, because not
enough words were provided to the table. Now 150 words are taken
from the model once an operation is filled in, but only the top
10 are shown. (I'm anticipating we'll want to add in a slider to
change the number of words; if so, we need to have the maximum
number of words available at the start.)

Commented out more outdated viz code in the Server Logic section.