Lots of HTML and CSS tweaks.
The top of the Viz. tab is now generated using an HTML template.
The opening paragraph of the Word Cloud description has been moved
to the top of the tab. All tab descriptions share a common format.

Also, incorporate feedback from @sydb on hover indicators for the
tab controls, and the padding of table headers relative to table
cells.